move test_opt to test_common
